Refactor: page.tsx 커스텀 훅 작성후 리팩토링
