+++ next.config.mjs
@@ -28,17 +28,33 @@ const nextConfig = {
     imageSizes: [20, 170, 310, 400, 580, 705],
     deviceSizes: [640, 768, 1024],
     remotePatterns: [
+      //supabase
       {
         protocol: 'https',
         hostname: process.env.NEXT_PUBLIC_HOST_NAME,
         port: '',
         pathname: '/storage/v1/object/public/**',
       },
+      //kakao
       {
         protocol: 'http',
         hostname: 't1.kakaocdn.net',
         port: '',
         pathname: '/account_images/**',
+      },
+      //google
+      {
+        protocol: 'https',
+        hostname: 'lh3.googleusercontent.com',
+        port: '',
+        pathname: '/a/**',
+      },
+      //naver
+      {
+        protocol: 'https',
+        hostname: 'lh3.googleusercontent.com',
+        port: '',
+        pathname: '/a/**',
       },
     ],
   },
