+++ src/language/ko.json
@@ -87,7 +87,7 @@
   "edittranslate-title-japanese": "한국어 번역 수정",
   "mypage-profile": "프로필",
   "mypage-nickname": "닉네임",
-  "mypage-profile-image": "프로피 이미지",
+  "mypage-profile-image": "프로필 이미지",
   "mypage-profile-email": "이메일",
   "mypage-profile-status": "상태",
   "mypage-profile-logout": "로그아웃",
